Adjust Test::Warnings usage to behave with fork()
The merged patch from Olaf didn't behave (for me at least) due to the
fork that happens when we actually start a server.  Instead of doing
something slightly funky and non-obvious to fix it, do the simple thing
and don't rely on Test::Warning's magic to get called.

Version 0.009 is required for :no_end_test support.